Completed Day 16 Part 1
Combined original idea of a brute force search through the graph with the idea of having a pre-filled list of paths between valves.  Now we perform the brute force search using the paths between valves with flow > 0 rather than the neighbors of each valve.  From the aoc subreddit I found the reasoning I was missing, which is: while there are 15 valves with flow > 0 and a brute force search across 15! possible paths is prohibitively expensive, the actual search space is much smaller, since we can simply stop going down a path, once all other valves cannot be reached in the remaining time