Completed Day 17 Part 2
Solution I found was to find a looping section of the rock stack, then multiply that by the number of times that loop would occur and add on the remaining bits at the start and end.  To find that loop, we drop shapes, recording each time the current combination of jet_index, shape_index, and floor*. Then when we reach a point where all three of those are the same, we know that the sections of rocks between those two points will loop.

The "floor" is a tuple of distances between the current top of the stack and the nearest rock below for each column. This is sufficient information, along with the jet and shape indices to tell us how the next shapes will land.  There is perhaps a more elegant way of finding this information, but this works fast enough for the time being.