Continuing work on Day 16 Part 2
Changed strategy to only ever consider either the human or elephant for each call of find_elephant_flow, and then call the function again WITHOUT adding to the flow or decreasing the time to handle cases where both finish opening their valve at the same time.  One possible concern is if there is only 1 valve left open at the end, the human will always try to reach it rather than the elephant. This might need to be fixed, or it might not, since we should explore symmetric paths with both the human and elephant. More investigation is needed on this.

This also may still need further optimizations to be fast enough. Currently taking a long time on slow hardware.