Completed Day 19 Part 2
Just increased the time to 32 and let it run.  This is really pushing up against reasonable time limits, but this still only took a couple minutes, so I'm going to call it good enough for now.

Another potential optimization to try later would be to find a way to store the states so that we could check if there is any state we have seen before that has less than or equal to the current state in all bots and resources, rather than just looking for states that are exactly the same.  This would probably speed things up fairly dramatically, but would require a more clever storage solution for the states than just a set