Completed Day 20 Parts 1 and 2
I suspect that there is a way to do each shift of number *k* in constant time, rather than O(k) time.  However, this version gets the correct answer and runs in less time than it would take me to figure out what that way might be, so we will call this good for now.