Completed Day 21 Part 2
This implementation is pretty sloppy, but effective and fast, and at this point I am just going to continue once I have a solution for the day.  I may try to come back later and clean this up.  The biggest thing right now would be to make both part 1 and 2 use the same underlying parsing of the input.  Right now they are totally separate