Day 24 Part 1: Began working on a_star
Created the basic skeleton of the a_star function.  Now the Valley class needs to be restructured so we can use a single instance of the class as the state in a_star.  This should be much faster than the DFS approach, but will take considerably more memory.  I may need to come up with a more clever way of storing the states, (THOUGHT! We may not need to save the positions of the Blizzards at all, since those should be easily derived from the initial positions and path length so far.  This would add some more computational pressure again, but greatly reduce the memory use)