Completed Day 24 Part 2
Pretty simple, just modified the search algorithm to take an arbitrary start state and goal, then run it three times going back and forth.  This is another one that is pushing up against the limit of reasonable running time, and there are surely still significant improvements that could be made, but they would take longer than just running the algorithm several times, so we will call this good for now.